requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
jinja2
python-multipart
